Fix frontend TypeScript build errors
