+++ frontend/src/components/TableView.tsx
@@ -1,6 +1,27 @@
-import React, { useState, useEffect } from 'react';
-import { X, Download, Filter, Search, ChevronLeft, ChevronRight, BarChart3, LineChart, PieChart, TrendingUp, Info } from 'lucide-react';
-import * as echarts from 'echarts';
+import React, { useState } from 'react';
+import { Search, ChevronLeft, ChevronRight } from 'lucide-react';
+
+const formatCellValue = (value: any, type?: string): string => {
+  if (value === null || value === undefined) return '-';
+
+  switch (type) {
+    case 'currency':
+      return `$${Number(value).toLocaleString('en-US', {
+        minimumFractionDigits: 2,
+        maximumFractionDigits: 2,
+      })}`;
+    case 'percentage':
+      return `${(Number(value) * 100).toFixed(2)}%`;
+    case 'number':
+      return Number(value).toLocaleString();
+    case 'date':
+      return new Date(value).toLocaleDateString('zh-CN');
+    case 'datetime':
+      return new Date(value).toLocaleString('zh-CN');
+    default:
+      return String(value);
+  }
+};
 
 
 interface TableViewProps {
